// src/util/types/Calendar.ts
/**
 * src/util/types/Calendar.ts
 *
 * Author: Ben Siebert <hello@ben-siebert.de>
 * Copyright: Copyright (c) 2018-2023 Ben Siebert. All rights reserved.
 * License: Project License
 * Created At: 13.11.2023
 *
 */

export type Calendar = {
  name: string;
  year: number;
  user: string;
  _id: string;
};